control-plane/pyproject.toml
[project]
name = "cloudbot-control-plane"
version = "1.0.0"
description = "Control plane for CloudBot cloud platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=44.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "boto3>=1.35.0",
    "websockets>=14.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.7.0",
    "aiofiles>=24.1.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
